+++ Messages/OMessages/buytransactionmsg.h
@@ -10,7 +10,7 @@ class BuyTransactionMsg : public OMessage
     qint32 m_amount;
     qint32 m_orderId;
 
-    qint32 length() const;
+    qint16 length() const;
 public:
 
     BuyTransactionMsg(qint32 orderId, qint32 amount);
