+++ Messages/OMessages/companyactive.cpp
@@ -20,7 +20,7 @@ void CompanyActiveResp::send(QIODevice *connection)
         << m_companyId;
 }
 
-qint32 CompanyActiveResp::length() const
+qint16 CompanyActiveResp::length() const
 {
     return -1;
 }
